backend/requirements.txt
Django
psycopg2-binary
boto3
django-storages
Pillow
djangorestframework
djangorestframework-simplejwt
django-filter
gunicorn
whitenoise
django-cors-headers
